+++ src/Adapter/IntendedDegreeProgram2/IntendedDegreeProgram2.mjs
@@ -1,6 +1,7 @@
+/** @typedef {import("../Combination/Combination.mjs").Combination} Combination */
 /** @typedef {import("./ChosenIntendedDegreeProgram2.mjs").ChosenIntendedDegreeProgram2} ChosenIntendedDegreeProgram2 */
 /** @typedef {import("../Subject/Subject.mjs").Subject} Subject */
 
 /**
- * @typedef {{subject: Subject, combination: string, values: ChosenIntendedDegreeProgram2 | null}} IntendedDegreeProgram2
+ * @typedef {{subject: Subject, combination: Combination, values: ChosenIntendedDegreeProgram2 | null}} IntendedDegreeProgram2
  */