+++ src/Adapter/IntendedDegreeProgram2/IntendedDegreeProgram2Element.mjs
@@ -1,4 +1,3 @@
-import { CssApi } from "../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs";
 import { ELEMENT_TAG_NAME_PREFIX } from "../Element/ELEMENT_TAG_NAME_PREFIX.mjs";
 import { FormElement } from "../Form/FormElement.mjs";
 import { PAGE_INTENDED_DEGREE_PROGRAM_2 } from "../Page/PAGE.mjs";
@@ -6,7 +5,9 @@ import { TitleElement } from "../Title/TitleElement.mjs";
 
 /** @typedef {import("../Post/backFunction.mjs").backFunction} backFunction */
 /** @typedef {import("./chosenIntendedDegreeProgram2Function.mjs").chosenIntendedDegreeProgram2Function} chosenIntendedDegreeProgram2Function */
+/** @typedef {import("../../Libs/flux-css-api/src/Adapter/Api/CssApi.mjs").CssApi} CssApi */
 /** @typedef {import("./IntendedDegreeProgram2.mjs").IntendedDegreeProgram2} IntendedDegreeProgram2 */
+/** @typedef {import("../../Service/Label/Port/LabelService.mjs").LabelService} LabelService */
 
 const __dirname = import.meta.url.substring(0, import.meta.url.lastIndexOf("/"));
 
@@ -32,20 +33,26 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
      */
     #intended_degree_program_2;
     /**
+     * @type {LabelService}
+     */
+    #label_service;
+    /**
      * @type {ShadowRoot}
      */
     #shadow;
 
     /**
      * @param {CssApi} css_api
+     * @param {LabelService} label_service
      * @param {IntendedDegreeProgram2} intended_degree_program_2
      * @param {chosenIntendedDegreeProgram2Function} chosen_intended_degree_program_2_function
      * @param {backFunction | null} back_function
      * @returns {IntendedDegreeProgram2Element}
      */
-    static new(css_api, intended_degree_program_2, chosen_intended_degree_program_2_function, back_function = null) {
+    static new(css_api, label_service, intended_degree_program_2, chosen_intended_degree_program_2_function, back_function = null) {
         return new this(
             css_api,
+            label_service,
             intended_degree_program_2,
             chosen_intended_degree_program_2_function,
             back_function
@@ -54,15 +61,17 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
 
     /**
      * @param {CssApi} css_api
+     * @param {LabelService} label_service
      * @param {IntendedDegreeProgram2} intended_degree_program_2
      * @param {chosenIntendedDegreeProgram2Function} chosen_intended_degree_program_2_function
      * @param {backFunction | null} back_function
      * @private
      */
-    constructor(css_api, intended_degree_program_2, chosen_intended_degree_program_2_function, back_function) {
+    constructor(css_api, label_service, intended_degree_program_2, chosen_intended_degree_program_2_function, back_function) {
         super();
 
         this.#css_api = css_api;
+        this.#label_service = label_service;
         this.#intended_degree_program_2 = intended_degree_program_2;
         this.#chosen_intended_degree_program_2_function = chosen_intended_degree_program_2_function;
         this.#back_function = back_function;
@@ -112,21 +121,25 @@ export class IntendedDegreeProgram2Element extends HTMLElement {
             "Subject",
             "readonly"
         );
-        subject_element.innerText = this.#intended_degree_program_2.subject.label;
-
-        const combination = this.#intended_degree_program_2.subject.combinations.find(_combination => _combination.id === this.#intended_degree_program_2.combination);
+        subject_element.innerText = this.#label_service.getSubjectLabel(
+            this.#intended_degree_program_2.subject
+        );
 
         const combination_element = this.#form_element.addInput(
             "Combination of Subjects",
             "readonly"
         );
-        combination_element.innerText = combination.label;
+        combination_element.innerText = this.#label_service.getCombinationLabel(
+            this.#intended_degree_program_2.combination
+        );
 
         const mandatory_element = this.#form_element.addInput(
             "Mandatory Subjects",
             "readonly"
         );
-        mandatory_element.innerText = combination.mandatory ?? "";
+        mandatory_element.innerText = this.#label_service.getMultipleMandatoryLabel(
+            this.#intended_degree_program_2.combination
+        );
 
         this.#form_element.addSubtitle(
             "If you have already studied the choosen subject and would like to continue in an advanced semester, please enter it here:"